dataplane: move files to dataplane/pkg/common
